Improve determining resize size when `set wrap`
Previously, when using the `set wrap` option, the area calculated to
resize the window to was the number of actual lines in the visual
selection, however the number of visibly selected lines can potentially
be more than this if some lines are wrapped.

This commit makes this behaviour more intuitive by calculating the
number of visible lines selected when this option is enabled and
resizing the window to that size instead, so the resized window should
always be the same as the visual selection rather than potentially
smaller.