Fix attempting to walk past last line in file
When calculating the number of visibly selected lines when `set wrap` is
enabled, if the last line in the file was selected we would previously
keep attempting to move the cursor beyond the end of the file, and so
loop indefinitely until cancelled by the user. We now check to see if
the cursor actually moved, and break out of the loop if it didn't.