Prevent previous window being overwritten
To determine if the current window has any window above or below it we
attempt to move to these windows and then back again. This overwrites
what the original previous window was when the visual-split command was
run, which can cause an unexpected window to be moved to if the user
then runs the `wincmd p` command.

This commit fixes this by remembering what this previous window
originally was and then resetting this if necessary, so the expected
window should then be moved to if `wincmd p` is run.